--- a/src/sagemaker/jumpstart/utils.py
+++ b/src/sagemaker/jumpstart/utils.py
@@ -1198,6 +1198,10 @@ def _deployment_config_lru_cache(_func=None, *, maxsize: int = 128, typed: bool 
 
             if f.cache_info().hits == 1:
                 print("******* Not From Cache ***********")
+                print(res)
+                print()
+                print(f.cache_info().misses)
+                print()
                 if isinstance(res, list):
                     for item in res:
                         if isinstance(