--- a/src/sagemaker/jumpstart/utils.py
+++ b/src/sagemaker/jumpstart/utils.py
@@ -1213,9 +1213,9 @@ def _deployment_config_lru_cache(_func=None, *, maxsize: int = 128, typed: bool 
                     keys = list(res.keys())
                     if "Instance Rate" not in keys[-1]:
                         f.cache_clear()
-                    elif len(res[keys[1]]) > len(res[keys[-1]]):
-                        del res[keys[-1]]
-                        f.cache_clear()
+                    # elif len(res[keys[1]]) > len(res[keys[-1]]):
+                    #     del res[keys[-1]]
+                    #     f.cache_clear()
             else:
                 print("******* From Cache ***********")
                 print(res)